+++ mayan/apps/storage/dependencies.py
@@ -1,16 +1,16 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='boto3', version_string='==1.33.7'
+    module=__name__, name='boto3', version_string='==1.34.105'
 )
 PythonDependency(
-    module=__name__, name='django-storages', version_string='==1.14.2'
+    module=__name__, name='django-storages', version_string='==1.14.3'
 )
 PythonDependency(
     module=__name__, name='extract-msg', version_string='==0.48.5'
 )
 PythonDependency(
-    module=__name__, name='google-cloud-storage', version_string='==2.15.0'
+    module=__name__, name='google-cloud-storage', version_string='==2.16.0'
 )
 PythonDependency(
     module=__name__, name='pycryptodome', version_string='==3.20.0'